init creates the default component files #371
