+++ cmd/init.go
@@ -24,6 +24,7 @@ var InitCmd = &cobra.Command{
 	PreRun: func(cmd *cobra.Command, args []string) {
 		viper.BindPFlag("network", cmd.Flags().Lookup("network"))
 		viper.BindPFlag("install-path", cmd.Flags().Lookup("install-path"))
+		viper.BindPFlag("redis-host", cmd.Flags().Lookup("redis-host"))
 	},
 	Run: func(cmd *cobra.Command, args []string) {
 		print.PendingStatusEvent(os.Stdout, "Making the jump to hyperspace...")
@@ -40,7 +41,8 @@ var InitCmd = &cobra.Command{
 		} else {
 			dockerNetwork := viper.GetString("network")
 			standalone.Uninstall(true, dockerNetwork)
-			err := standalone.Init(runtimeVersion, dockerNetwork, installLocation)
+			redisHost := viper.GetString("redis-host")
+			err := standalone.Init(runtimeVersion, dockerNetwork, installLocation, redisHost)
 			if err != nil {
 				print.FailureStatusEvent(os.Stdout, err.Error())
 				return
@@ -55,6 +57,7 @@ func init() {
 	InitCmd.Flags().StringVarP(&runtimeVersion, "runtime-version", "", "latest", "The version of the Dapr runtime to install. for example: v0.1.0")
 	InitCmd.Flags().String("network", "", "The Docker network on which to deploy the Dapr runtime")
 	InitCmd.Flags().String("install-path", "", "The optional location to install Dapr to.  The default is /usr/local/bin for Linux/Mac and C:\\dapr for Windows")
+	InitCmd.Flags().String("redis-host", "localhost", "The host on which the Redis service resides")
 
 	RootCmd.AddCommand(InitCmd)
 }
