+++ docs/reference/dapr-init.md
@@ -18,3 +18,4 @@ dapr init [flags]
 | `--help`, `-h` | | | Help for init |
 | `--network` | `DAPR_NETWORK` | | The Docker network on which to deploy the Dapr runtime |
 | `--runtime-version` | | `latest` | The version of the Dapr runtime to install, for example: `v0.1.0-alpha` |
+| `--redis-host` | `DAPR_REDIS_HOST` | `localhost` | The host on which the Redis service resides |
