+++ docs/reference/dapr-run.md
@@ -28,4 +28,3 @@ dapr run [flags] [command]
 | `--port`, `-p` | | `-1` | The HTTP port for Dapr to listen on |
 | `--profile-port` | | `-1` | The port for the profile server to listen on |
 | `--protocol` | | `http` | Tells Dapr to use HTTP or gRPC to talk to the app. Valid values are: `http` or `grpc` |
-| `--redis-host` | `DAPR_REDIS_HOST` | `localhost` | The host on which the Redis service resides |
